pyproject.toml
[project]
name = "ai-app"
version = "0.1.0"
requires-python = ">=3.7"
readme = "README.md"
authors = [{ name = "Team", email = "team@team.com" }]

dependencies = [
  # Api server libraries
  "fastapi",
  "typer",
  "uvicorn",
  # Database libraries
  "alembic",
  "duckdb",
  "pgvector",
  "psycopg[binary]",
  "sqlalchemy",
  # Project libraries
  "beautifulsoup4",
  "bs4",
  "duckduckgo-search",
  "exa_py",
  "matplotlib",
  "nest_asyncio",
  "openai",
  "openai",
  "pandas",
  "pillow",
  "pypdf",
  "streamlit",
  "tiktoken",
  "types-beautifulsoup4",
  "types-Pillow",
  "yfinance",
  # Type checking
  "mypy",
  # Linting and Formatting
  "ruff",
  # phidata
  "phidata[aws]==2.4.15",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]

# Change this value to use a different directory for the phidata workspace.
# [tool.phidata]
# workspace = "workspace"

[tool.ruff]
line-length = 110
extend-include = ["*.ipynb"]
exclude = ["aienv*", ".venv*"]
[tool.ruff.lint.per-file-ignores]
# Ignore `F401` (import violations) in all `__init__.py` files
"__init__.py" = ["F401"]
"*.ipynb" = ["E402"]

[tool.mypy]
check_untyped_defs = true
no_implicit_optional = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
exclude = ["aienv*", ".venv*"]

[[tool.mypy.overrides]]
module = ["pgvector.*", "setuptools.*", "nest_asyncio.*"]
ignore_missing_imports = true
